+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:theme="@style/Theme.VotingApp"
         tools:targetApi="31">
         <activity
-            android:name=".LoginPage"
+            android:name=".SignUpPage"
             android:exported="true"
-            android:parentActivityName=".LoginPage">
+            android:parentActivityName=".SignUpPage">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
