+++ app/src/main/res/drawable/ic_email_foreground.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="18"
+    android:viewportHeight="18"
+    android:tint="#000000">
+  <group android:scaleX="0.78"
+      android:scaleY="0.78"
+      android:translateX="0.00"
+      android:translateY="0.00">
+    <path
+        android:fillColor="@android:color/black"
+        android:pathData="M20,4L4,4c-1.1,0 -1.99,0.9 -1.99,2L2,18c0,1.1 0.9,2 2,2h16c1.1,0 2,-0.9 2,-2L22,6c0,-1.1 -0.9,-2 -2,-2zM20,8l-8,5 -8,-5L4,6l8,5 8,-5v2z"/>
+  </group>
+</vector>
