+++ app/src/main/res/drawable/ic_name_foreground.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="18"
+    android:viewportHeight="18"
+    android:tint="#000000">
+  <group android:scaleX="0.78"
+      android:scaleY="0.78"
+      android:translateX="0.00"
+      android:translateY="0.00">
+    <path
+        android:fillColor="#000000"
+        android:pathData="M12,12c2.21,0 4,-1.79 4,-4s-1.79,-4 -4,-4 -4,1.79 -4,4 1.79,4 4,4zM12,14c-2.67,0 -8,1.34 -8,4v2h16v-2c0,-2.66 -5.33,-4 -8,-4z"/>
+  </group>
+</vector>
